backend/pyproject.toml
# Python configuration file for building, dependencies, and tooling.
#
# For more information, visit
# https://snarky.ca/what-the-heck-is-pyproject-toml/.

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
extend-exclude = "^/template"
line-length = 80

[tool.coverage.report]
exclude_lines = [
    "if sys.version_info.*",
    "pragma: no cover",
    "raise NotImplementedError",
]
fail_under = 100
show_missing = true

[tool.coverage.run]
branch = true
command_line = "-m pytest"
omit = []
source = ["acronyms"]

[tool.poetry]
authors = ["Macklan Weinstein <macklan.weinstein@gmail.com>"]
description = "Application for saving acronyms"
documentation = "https://scruffaluff.gitlab.io/acronyms"
homepage = "https://gitlab.com/scruffaluff/acronyms"
keywords = []
license = "MIT"
name = "acronyms"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
fastapi = "^0.74.0"
google-cloud-bigquery = "^2.0.0"
psycopg = {extras = ["binary", "pool"], version = "^3.0.9"}
psycopg2-binary = "^2.9.3"
pybigquery = "^0.10.0"
pydantic = "^1.9.0"
python = "^3.9.0,<3.10.0"
SQLAlchemy = "^1.0.0"
typer = "^0.4.0"
uvicorn = {extras = ["standard"], version = "^0.17.5"}

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black = "^21.12b0"
coverage = { version = "^6.3.0", extras = ["toml"] }
flake8 = "^4.0.0"
flake8-bugbear = "^22.1.11"
flake8-docstrings = "^1.6.0"
flake8-import-order = "^0.18.0"
mccabe = "^0.6.0"
mypy = "^0.931"
pdbpp = "^0.10.0"
py-spy = "^0.3.0"
pytest = "^7.0.0"
pytest-cov = "^3.0.0"
pytest-mock = "^3.7.0"
pytest-postgresql = "^4.1.0"
sqlalchemy-stubs = "^0.4"
tinyquery = "^1.2.1"
toml = "^0.10.0"
tox = "^3.0.0"

[tool.poetry.scripts]
acronyms = "acronyms.__main__:app"

[tool.pytest.ini_options]
addopts = "-sv"
markers = [
    "functional: mark as functional test.",
    "integration: mark as integration test.",
    "unit: mark as unit test.",
]
python_classes = "Test*"
python_files = [
    "__init__.py",
    "test_*.py",
]
python_functions = ["test_*"]
testpaths = ["tests"]
